requirements.txt
pytesseract
Pillow
pyautogui
pywinctl
